# GGRC-5776 Improve performance of "setColumnsConfiguration" method