@romamatusevich 2018-08-17
Memory leak issue with controls was mostly fixed by other performance related fixes. It's not possible to increase memory usage to 1 GB and the number of DOM Nodes to 240 thousands.